+++ crates/iceberg/src/arrow/mod.rs
@@ -29,6 +29,7 @@ pub(crate) mod delete_filter;
 
 mod int96;
 mod reader;
+pub(crate) mod timestamp_tz;
 /// RecordBatch projection utilities
 pub mod record_batch_projector;
 pub(crate) mod record_batch_transformer;
